Removed OAuthServerBundle which was confusing because of FOSOAuthServerBundle already installed. OAuth Model classes have been moved to BileMo/AppBundle/Entity/OAuth dir, createClient command have been moved to BileMo/AppBundle/Command dir
