+++ src/BileMo/AppBundle/Command/CreateClientCommand.php
@@ -1,9 +1,10 @@
 <?php
-namespace OAuthServerBundle\Command;
+namespace BileMo\AppBundle\Command;
 use Symfony\Bundle\FrameworkBundle\Command\ContainerAwareCommand;
 use Symfony\Component\Console\Input\InputInterface;
 use Symfony\Component\Console\Input\InputOption;
 use Symfony\Component\Console\Output\OutputInterface;
+
 class CreateClientCommand extends ContainerAwareCommand
 {
     protected function configure()
