+++ src/BileMo/AppBundle/Entity/OAuth/AccessToken.php
@@ -1,13 +1,13 @@
 <?php
-namespace OAuthServerBundle\Entity;
+namespace BileMo\AppBundle\Entity\OAuth;
 
-use FOS\OAuthServerBundle\Entity\AuthCode as BaseAuthCode;
+use FOS\OAuthServerBundle\Entity\AccessToken as BaseAccessToken;
 use Doctrine\ORM\Mapping as ORM;
 
 /**
  * @ORM\Entity
  */
-class AuthCode extends BaseAuthCode
+class AccessToken extends BaseAccessToken
 {
     /**
      * @ORM\Id
