--- src/OAuthServerBundle/Controller/DefaultController.php
@@ -1,13 +1,0 @@
-<?php
-
-namespace OAuthServerBundle\Controller;
-
-use Symfony\Bundle\FrameworkBundle\Controller\Controller;
-
-class DefaultController extends Controller
-{
-    public function indexAction()
-    {
-        return $this->render('OAuthServerBundle:Default:index.html.twig');
-    }
-}
