--- src/OAuthServerBundle/OAuthServerBundle.php
@@ -1,9 +1,0 @@
-<?php
-
-namespace OAuthServerBundle;
-
-use Symfony\Component\HttpKernel\Bundle\Bundle;
-
-class OAuthServerBundle extends Bundle
-{
-}
