--- src/OAuthServerBundle/Resources/config/services.yml
@@ -1,4 +1,0 @@
-services:
-#    o_auth_server.example:
-#        class: OAuthServerBundle\Example
-#        arguments: ["@service_id", "plain_value", "%parameter%"]
